Include all points in spiral. Explanation why this changed below.
Before I couldn't do this because the first insertion onto the map is at (0,0),
overwriting the value with 0 (because all (0,0)'s neighbors are zero). Once I
changed how the sample method to include (0,0), I could then sample a point
when creating the point, so excluding (0,0) was no longer needed.

I think it's nice to include (0,0) from an aesthetic point of view.